+++ src/VaultSharp/V1/SecretsEngines/Consul/Models/ConsulTokenType.cs
@@ -1,0 +1,13 @@
+﻿
+using Newtonsoft.Json;
+using Newtonsoft.Json.Converters;
+
+namespace VaultSharp.V1.SecretsEngines.Consul.Models
+{
+    [JsonConverter(typeof(StringEnumConverter))]
+    public enum ConsulTokenType
+    {
+        client,
+        management
+    }
+}